use new component everywhere
