+++ frontend/src/lib/components/AppEditorLink/AppEditorLink.tsx
@@ -1,10 +1,9 @@
 import React, { useState } from 'react'
 import { useValues } from 'kea'
-
-import { EditAppUrls } from './EditAppUrls'
-import { appEditorUrl } from './utils'
 import { teamLogic } from 'scenes/teamLogic'
 import { Modal, Button } from 'antd'
+import { AuthorizedUrlsTable } from 'scenes/toolbar-launch/AuthorizedUrlsTable'
+import { appEditorUrl } from 'scenes/toolbar-launch/authorizedUrlsLogic'
 
 export function AppEditorLink({
     actionId,
@@ -41,7 +40,7 @@ export function AppEditorLink({
                 footer={<Button onClick={() => setModalOpen(false)}>Close</Button>}
                 onCancel={() => setModalOpen(false)}
             >
-                <EditAppUrls actionId={actionId} allowNavigation />
+                <AuthorizedUrlsTable actionId={actionId} pageKey="app-editor-link" />
             </Modal>
         </>
     )
